address feedback - create hocs folder
